+++ src/components/transactionCard/TransactionsList.tsx
@@ -12,7 +12,6 @@ interface ITransactionList {
 const TransactionsList = ({ data, listType }: ITransactionList) => {
     const queryClient = useQueryClient();
 
-    console.log(data);
     const mutation = useMutation({
         mutationFn: deleteTransaction,
         onSuccess: () => {
@@ -39,5 +38,4 @@ const TransactionsList = ({ data, listType }: ITransactionList) => {
         </>
     )
 }
-
 export default TransactionsList;
