+++ src/config/typeTransactions.ts
@@ -1,11 +1,7 @@
-interface ITypesTransaction {
-  EXPENSE: number;
-  INCOME: number;
-}
+import { ITypesTransaction } from "../interface/ITypesTransaction";
 
 const TYPES_TRANSACTION: ITypesTransaction = {
   EXPENSE: 1,
   INCOME: 2,
 };
-
 export default TYPES_TRANSACTION;
