+++ src/helpers/calculateStats.ts
@@ -6,7 +6,7 @@ type CalcStatsFn = {
   incomes: Array<ITransaction>;
 };
 
-export default function calcStats(fn: CalcStatsFn): string[] {
+export default function calculateStats(fn: CalcStatsFn): string[] {
   const totalExpenses: number = fn.expenses.reduce((acc: number, item: ITransaction) => acc + item.amount, 0);
   const totalIncomes: number = fn.incomes.reduce((acc: number, item: ITransaction) => acc + item.amount, 0);
