--- src/helpers/dateFormat.ts
@@ -1,4 +1,0 @@
-export function formatDate(date: string): string {
-    const options = { year: 'numeric', month: 'short', day: 'numeric' } ;
-    return new Date(date).toLocaleDateString('en-US', options);
-}
