--- src/helpers/sortParams.ts
@@ -1,1 +1,0 @@
-export default function getSortParams(params: string) {}
