+++ src/interface/ITypesTransaction.ts
@@ -1,0 +1,4 @@
+export interface ITypesTransaction {
+  EXPENSE: number;
+  INCOME: number;
+}
