--- src/layout/loader.ts
@@ -1,8 +1,0 @@
-import supabase from "../services/supabase";
-
-export default async function loadData() {
-  const data = supabase.from("category").select();
-  console.log(data);
-
-  return data;
-}
